+++ test/doc_test.exs
@@ -2,6 +2,5 @@ defmodule DocTest do
   use ExUnit.Case, async: true
 
   doctest Comeonin
-  doctest Comeonin.Tools
 
 end